Project Updates
Speed up the animations in level one, added a barrier to death scenes, fixed a bug that only played the player one death scene, made the players In goofy poses in the death scenes, added collision sound when players collide, Replaced the keemstar scream for the player one with a women scream. Im not gonna really worry about the movement I think it looks fine. Added Death Scenes for level 3, Fixed a bug with shooting